requirements.txt
langchain_openai
fitz==0.0.1.dev2
pymupdf==1.24.10
python-dotenv==1.0.1
langchain==0.3.0
yacs==0.1.8
transformers==4.44.2
torch==2.4.1
pillow==10.4.0
torchvision==0.19.1
fastapi 
uvicorn 
python-jose 
passlib
cryptography
bcrypt
python-multipart
pymongo
pdf2image